changing to size_t and signed unsigned

+++ include/_mem.h
@@ -7,10 +7,10 @@ BEGIN_DECLS
 
 extern const Except_T Mem_Failed;
 
-extern void *_Mem_alloc (long nbytes, const char *file, int line);
-extern void *_Mem_calloc(long count, long nbytes, const char *file, int line);
+extern void *_Mem_alloc (size_t nbytes, const char *file, int line);
+extern void *_Mem_calloc(size_t count, size_t nbytes, const char *file, int line);
 extern void _Mem_free(void *ptr, const char *file, int line);
-extern void *_Mem_realloc(void *ptr, long nbytes, const char *file, int line);
+extern void *_Mem_realloc(void *ptr, size_t nbytes, const char *file, int line);
 extern void _Mem_print_stats();
 
 END_DECLS
